requirements.txt
django==4.2.11 
djangorestframework==3.14.0 
psycopg2-binary==2.9.9 
drf-spectacular==0.27.1 
django-filter==24.2